Huge update.  VUE app now has multiple map views.  We now have a ToolWindow, and two working examples - inspector & panner. (Panner is display only at the moment).  Multiple item selection on the map is partially implemented (display only -- can't modify an entire selection yet).  Also threw in some place holder icon images.
git-svn-id: https://tuftsvue.svn.sourceforge.net/svnroot/tuftsvue/trunk@13 d66e5487-c4fa-4dda-8c8e-ca23240363c0

+++ VUE2/src/tufts/vue/VueUtil.java
@@ -7,14 +7,14 @@ public class VueUtil
     private static boolean WindowsPlatform = false;
     private static boolean MacPlatform = false;
     private static boolean UnixPlatform = false;
-    private static double javaVersion = 1.0f;
+    private static float javaVersion = 1.0f;
 
     static {
         String osName = System.getProperty("os.name");
         String javaSpec = System.getProperty("java.specification.version");
 
         try {
-            javaVersion = Double.parseDouble(javaSpec);
+            javaVersion = Float.parseFloat(javaSpec);
             System.out.println("Java Version: " + javaVersion);
         } catch (Exception e) {
             System.err.println("Couldn't parse java.specifcaion.version: [" + javaSpec + "]");
